+++ examples/siro_sandbox/sandbox_app.py
@@ -9,12 +9,10 @@ See README.md in this directory.
 """
 
 import ctypes
-import math
 
 # must call this before importing habitat or magnum! avoids EGL_BAD_ACCESS error on some platforms
 import sys
 from enum import Enum
-from math import radians
 
 flags = sys.getdlopenflags()
 sys.setdlopenflags(flags | ctypes.RTLD_GLOBAL)
@@ -26,6 +24,7 @@ from typing import Any, List, Tuple
 import magnum as mn
 import numpy as np
 from controllers import ControllerHelper, GuiHumanoidController
+from hitl_tutorial import TutorialStage, generate_tutorial
 from magnum.platform.glfw import Application
 
 import habitat
@@ -61,64 +60,8 @@ def requires_habitat_sim_with_bullet(callable_):
 
 
 class SandboxState(Enum):
-    SIMULATION = 1
+    CONTROLLING_AGENT = 1
     TUTORIAL = 2
-
-
-class TutorialStage:
-    _prev_lookat: Tuple[mn.Vector3, mn.Vector3]
-    _next_lookat: Tuple[mn.Vector3, mn.Vector3]
-    _transition_duration: float
-    _stage_duration: float
-    _elapsed_time: float
-
-    def __init__(
-        self,
-        stage_duration: float,
-        next_lookat: Tuple[mn.Vector3, mn.Vector3],
-        prev_lookat: Tuple[mn.Vector3, mn.Vector3] = None,
-        transition_duration: float = 0.0,
-    ) -> None:
-        self._transition_duration = transition_duration
-        self._stage_duration = stage_duration
-        self._prev_lookat = prev_lookat
-        self._next_lookat = next_lookat
-        self._elapsed_time = 0.0
-
-    def update(self, dt: float) -> None:
-        self._elapsed_time += dt
-
-    def get_look_at_matrix(self) -> mn.Matrix4:
-        # If there's no transition, return the next view
-        assert self._next_lookat
-        if not self._prev_lookat or self._transition_duration <= 0.0:
-            return mn.Matrix4.look_at(
-                self._next_lookat[0], self._next_lookat[1], mn.Vector3(0, 1, 0)
-            )
-        # Interpolate camera look-ats
-        t: float = (
-            _ease_fn_in_out_quat(
-                min(
-                    self._elapsed_time / self._transition_duration,
-                    1.0,
-                )
-            )
-            if self._transition_duration > 0.0
-            else 1.0
-        )
-        look_at: List[mn.Vector3] = []
-        for i in range(2):  # Only interpolate eye and target vectors
-            look_at.append(
-                mn.math.lerp(
-                    self._prev_lookat[i],
-                    self._next_lookat[i],
-                    t,
-                )
-            )
-        return mn.Matrix4.look_at(look_at[0], look_at[1], mn.Vector3(0, 1, 0))
-
-    def is_stage_completed(self) -> bool:
-        return self._elapsed_time >= self._stage_duration
 
 
 @requires_habitat_sim_with_bullet
@@ -188,10 +131,16 @@ class SandboxDriver(GuiAppDriver):
         self._sandbox_state = (
             SandboxState.TUTORIAL
             if args.show_tutorial
-            else SandboxState.SIMULATION
+            else SandboxState.CONTROLLING_AGENT
         )
         self._tutorial_stages: List[TutorialStage] = (
-            self._generate_tutorial() if args.show_tutorial else None
+            generate_tutorial(
+                self.get_sim(),
+                self.ctrl_helper.get_gui_controlled_agent_index(),
+                self._create_camera_lookat(),
+            )
+            if args.show_tutorial
+            else None
         )
         self._tutorial_stage_index: int = 0
 
@@ -670,7 +619,7 @@ class SandboxDriver(GuiAppDriver):
 
         return (lookat + offset.normalized() * self.cam_zoom_dist, lookat)
 
-    def _sim_update_simulation(self, dt: float):
+    def _sim_update_controlling_agent(self, dt: float):
         if isinstance(self.gui_agent_ctrl, GuiHumanoidController):
             (
                 walk_dir,
@@ -716,84 +665,6 @@ class SandboxDriver(GuiAppDriver):
             lookat[0], lookat[1], mn.Vector3(0, 1, 0)
         )
 
-    def _generate_tutorial(self) -> List[TutorialStage]:
-        tutorial_stages: List[TutorialStage] = []
-
-        sim = self.get_sim()
-        camera_fov_deg = 90  # TODO: Get the actual FOV
-
-        # Scene overview
-        scene_root_node = sim.get_active_scene_graph().get_root_node()
-        scene_target_bb: mn.Range3D = scene_root_node.cumulative_bb
-        scene_top_down_lookat = _lookat_bounding_box_top_down(
-            camera_fov_deg, scene_target_bb
-        )
-        tutorial_stages.append(
-            TutorialStage(
-                stage_duration=5.0, next_lookat=scene_top_down_lookat
-            )
-        )
-
-        # Show all the targets
-        pathfinder = sim.pathfinder
-        idxs, goal_pos = sim.get_targets()
-        scene_positions = sim.get_scene_pos()
-        target_positions = scene_positions[idxs]
-        for target_pos in target_positions:
-            next_lookat = _lookat_point_from_closest_navmesh_pos(
-                mn.Vector3(target_pos), 0.75, 1.5, pathfinder
-            )
-            tutorial_stages.append(
-                TutorialStage(
-                    stage_duration=3.0,
-                    transition_duration=2.0,
-                    prev_lookat=tutorial_stages[
-                        len(tutorial_stages) - 1
-                    ]._next_lookat,
-                    next_lookat=next_lookat,
-                )
-            )
-
-        # Controlled agent focus
-        agent_idx = self.ctrl_helper.get_gui_controlled_agent_index()
-        assert agent_idx is not None
-        art_obj = (
-            self.get_sim().agents_mgr[agent_idx].articulated_agent.sim_obj
-        )
-        agent_root_node = art_obj.get_link_scene_node(
-            -1
-        )  # Root link always has index -1
-        target_bb = mn.Range3D.from_center(
-            mn.Vector3(agent_root_node.absolute_translation),
-            mn.Vector3(1.0, 1.0, 1.0),
-        )  # Assume 2x2x2m bounding box
-        agent_lookat = _lookat_bounding_box_top_down(camera_fov_deg, target_bb)
-
-        tutorial_stages.append(
-            TutorialStage(
-                stage_duration=2.0,
-                transition_duration=1.0,
-                prev_lookat=tutorial_stages[
-                    len(tutorial_stages) - 1
-                ]._next_lookat,
-                next_lookat=agent_lookat,
-            )
-        )
-
-        # Transition from the avatar view to simulated view (e.g. first person)
-        tutorial_stages.append(
-            TutorialStage(
-                stage_duration=1.0,
-                transition_duration=1.0,
-                prev_lookat=tutorial_stages[
-                    len(tutorial_stages) - 1
-                ]._next_lookat,
-                next_lookat=self._create_camera_lookat(),
-            )
-        )
-
-        return tutorial_stages
-
     def _sim_update_tutorial(self, dt: float):
         # Keyframes are saved by RearrangeSim when stepping the environment.
         # Because the environment is not stepped in the tutorial, we need to save keyframes manually for replay rendering to work.
@@ -807,7 +678,7 @@ class SandboxDriver(GuiAppDriver):
         if tutorial_stage.is_stage_completed():
             self._tutorial_stage_index += 1
             if self._tutorial_stage_index >= len(self._tutorial_stages):
-                self._sandbox_state = SandboxState.SIMULATION
+                self._sandbox_state = SandboxState.CONTROLLING_AGENT
 
     def sim_update(self, dt):
         # todo: pipe end_play somewhere
@@ -831,8 +702,8 @@ class SandboxDriver(GuiAppDriver):
                 not self.env._sim.navmesh_visualization  # type: ignore
             )
 
-        if self._sandbox_state == SandboxState.SIMULATION:
-            self._sim_update_simulation(dt)
+        if self._sandbox_state == SandboxState.CONTROLLING_AGENT:
+            self._sim_update_controlling_agent(dt)
         else:
             self._sim_update_tutorial(dt)
 
@@ -872,7 +743,7 @@ class SandboxDriver(GuiAppDriver):
             np.flipud(image) for image in debug_images
         ]
 
-        if self._sandbox_state == SandboxState.SIMULATION:
+        if self._sandbox_state == SandboxState.CONTROLLING_AGENT:
             self._update_text()
 
         return post_sim_update_dict
@@ -901,69 +772,6 @@ def _evaluate_cubic_bezier(ctrl_pts, t):
         result += weights[i] * ctrl_pts[i]
 
     return result
-
-
-def _ease_fn_in_out_quat(t: float):
-    if t < 0.5:
-        return 16 * t * pow(t, 4)
-    else:
-        return 1 - pow(-2 * t + 2, 4) / 2
-
-
-def _lookat_bounding_box_top_down(
-    camera_fov_deg: float, target_bb: mn.Range3D
-) -> Tuple[mn.Vector3, mn.Vector3]:
-    r"""
-    Creates look-at vectors for a top-down camera such as the entire 'target_bb' bounding box is visible.
-    """
-    camera_fov_rad = radians(camera_fov_deg)
-    target_dimension = max(target_bb.size_x(), target_bb.size_z())
-    camera_position = mn.Vector3(
-        target_bb.center_x(),
-        target_bb.center_y()
-        + abs(target_dimension / math.sin(camera_fov_rad / 2)),
-        target_bb.center_z(),
-    )
-    epsilon = 0.0001  # Because of gimbal lock, we apply an epsilon bias to force a camera orientation
-    return (
-        camera_position,
-        target_bb.center() + mn.Vector3(0.0, 0.0, epsilon),
-    )
-
-
-def _lookat_point_from_closest_navmesh_pos(
-    target: mn.Vector3,
-    distance_from_target: float,
-    viewpoint_height: float,
-    pathfinder: habitat_sim.PathFinder,
-) -> Tuple[mn.Vector3, mn.Vector3]:
-    r"""
-    Creates look-at vectors for a viewing a point from a nearby navigable point (with a height offset).
-    Helps finding a viewpoint that is not occluded by a wall or other obstacle.
-    """
-    # Look up for a camera position by sampling radially around the target for a navigable position.
-    navigable_point = None
-    sample_count = 8
-    for i in range(sample_count):
-        radial_angle = i * 2.0 * math.pi / float(sample_count)
-        dist_x = math.sin(radial_angle) * distance_from_target
-        dist_z = math.cos(radial_angle) * distance_from_target
-        candidate = mn.Vector3(target.x + dist_x, target.y, target.z + dist_z)
-        if pathfinder.is_navigable(candidate, 3.0):
-            navigable_point = candidate
-
-    if navigable_point == None:
-        # Fallback to a default point
-        navigable_point = mn.Vector3(
-            target.x + distance_from_target, target.y, target.z
-        )
-
-    camera_position = mn.Vector3(
-        navigable_point.x,
-        navigable_point.y + viewpoint_height,
-        navigable_point.z,
-    )
-    return (camera_position, target)
 
 
 def parse_debug_third_person(args, framebuffer_size):
